requirements.txt
streamlit==1.29.0
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.0
requests==2.31.0
